[FEAT] 방문 접속자를 서버가 아닌 클라이언트에서 이벤트를 전송하게 변경 (#1097)
* refactor: 유저 행동 추적 기반 컴포넌트 및 훅 폴더 변경

* feat: 로그인한 유저가 아닐 때는 식별 못하게 변경

* feat: ScreenLogger, client component로 변경

* feat: logout시 세션 초기화
