+++ frontend/techpick/src/app/(unsigned)/share/[uuid]/page.tsx
@@ -4,8 +4,8 @@ import { FolderContentLayout } from '@/components/FolderContentLayout';
 import { MobileEmptyPickRecordImage } from '@/components/MobileEmptyPickRecordImage';
 import { PickContentLayout } from '@/components/PickContentLayout';
 import { PickRecordHeader } from '@/components/PickRecord/PickRecordHeader';
-import { ScreenLogger } from '@/components/ScreenLogger';
 import { ROUTES } from '@/constants/route';
+import { ScreenLogger } from '@/libs/@eventlog/ScreenLogger';
 import { isLoginUser } from '@/utils/isLoginUser';
 import { FolderOpenIcon } from 'lucide-react';
 import type { Metadata, ResolvingMetadata } from 'next';
