+++ frontend/techpick/src/app/layout.tsx
@@ -2,9 +2,9 @@ import { QueryProvider } from '@/providers/QueryProvider';
 import { Noto_Sans_KR } from 'next/font/google';
 import '@/styles/reset.css';
 import { PORTAL_CONTAINER_ID } from '@/constants/portalContainerId';
+import { UserIdentifyProvider } from '@/libs/@eventlog/UserIdentifyProvider';
 import { ThemeProvider } from '@/providers/ThemeProvider';
 import { ToastProvider } from '@/providers/ToastProvider';
-import { UserIdentifyProvider } from '@/providers/UserIdentifyProvider';
 import { getUserIdForServer } from '@/utils/getUserIdForServer';
 import type { Metadata } from 'next';
 import { preventOverscrollBehaviorY } from './layout.css';
